+++ R/data-st_fips_map.R
@@ -6,8 +6,8 @@
 #' \describe{
 #'   \item{name}{character, name of territory}
 #'   \item{abbrev}{character, abbreviation}
-#'   \item{fips}{character, FIPS code}
-#'   \item{nchs}{character, NCHS state code}
+#'   \item{fips}{numeric, FIPS state code (zero-pad to 2 digits for a FIPS string)}
+#'   \item{nchs}{numeric, NCHS state code (zero-pad to 2 digits)}
 #' }
 #' @source \url{https://en.wikipedia.org/wiki/Federal_Information_Processing_Standard_state_code}
 #' @keywords datasets